% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCD-Implement.R
\name{BTSBM}
\alias{BTSBM}
\title{Binary Tree Stochastic Block Models (BTSBM)}
\usage{
BTSBM(n, d, a.seq, lambda, alpha = NULL, N = 1)
}
\arguments{
\item{n}{dimension of the network}

\item{d}{number of layers until leaves (excluding the root)}

\item{a.seq}{sequence a_r}

\item{lambda}{average node degree, only used when alpha is not provided.}

\item{alpha}{the common scaling of the a_r sequence, so at the end, essentially the a_r sequence is a.seq*alpha}

\item{N}{number of networks one wants to generate from the same model}
}
\value{
Output a \code{HCD} object
\describe{
\item{A.list}{list of simulated networks}
\item{B}{clusters' probability matrix}
\item{label}{cluster label}
\item{P}{nodes' probability matrix}
\item{comm.sim.mat}{clusters' similarity matrix}
\item{node.sim.mat}{nodes' similarity matrix}
}
}
\description{
Simulated network from BTSBM.
}
\examples{
BTSBM(1024, 2, c(1, .3, .09), 50)
}
