% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citation}
\alias{citation}
\title{Citation network}
\format{
A matrix of citation network among 707 statisicians.
}
\usage{
citation
}
\description{
A dataset contatining the 3-core of the largest connected component of
the citation network by Ji and Jin (2016).
}
\keyword{datasets}
