% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hvsp.R
\name{Hvsp}
\alias{Hvsp}
\title{Hierarchical Vintage Sparse PCA (Hvsp)}
\usage{
Hvsp(A, stopping = "NB", n.min = 25, D = NULL, notree = TRUE, ...)
}
\arguments{
\item{A}{input adjacency matrix. Can be standard R matrix or dsCMatrix (or other type in package Matrix)}

\item{stopping}{stopping rule. "NB" (default) -- non-backtracking matrix spectrum, "ECV" -- edge cross-validation, "Fix"-- fixed D layers of partitioning
ECV is nonparametric rank evaluation by cross-validation -- a more generally applicable approach
without assuming SBM or its variants. Also available for weighted networks.
It is believed to be more robust than NB but less effective if the true model is close to BTSBM.
ECV is computationally much more intensive.}

\item{n.min}{the algorithm will stop splitting if the current size is <= 2*n.min.}

\item{D}{the number of layers to partition, if stopping=="Fix".}

\item{notree}{if TRUE, will not produce the data.tree object or the community similarity matrix. Only the cluster label and the tree path strings will be returned. This typically makes the runing faster.}

\item{...}{optional arguments for vsp()}
}
\value{
Output a \code{HCD} object
\describe{
\item{results}{
\describe{
\item{labels}{clustering label}
\item{ncl}{number of clusters}
\item{cluster.tree}{cluster tree from data.tree}
\item{P}{estimated probability matrix}
\item{node.bin.sim.mat}{nodes' similarity matrix}
\item{comm.bin.sim.mat}{clusters' similarity matrix}
\item{tree.path}{nodes' tree path}
}
}
\item{cent}{centrality/importance score for vsp}
}
}
\description{
The hierarchical vintage sparse PCA (Hvsp) that combines
vintage sparse PCA by Rohe and Zeng (2022) and
hierarchical community detection (HCD) by Li et al. (2020).
}
\examples{
tt <- BTSBM(1024, 2, c(1, .3, .09), 50)
A <- tt$A.list[[1]]
Hvsp(A)

data(citation)
Hvsp(citation, method="SS", stopping="NB", notree=F)
}
